model-api/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = ""
authors = [
    {name = "vishwajithkasthoori",email = "sandarukasthoori@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
python = ">=3.13,<3.14"
dependencies = [
    "torch (==2.8.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "flask (>=3.1.2,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[tool.poetry]
packages = [{include = "model", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
